Move logging tests to dedicated test file
